Refactor Categories controller and bussines to handle page range errors. Relocate PageList class.
